+++ flask_unchained/bundles/graphene/object_types.py
@@ -1,12 +1,34 @@
 import graphene
 
 from flask_unchained import unchained
+from flask_unchained.bundles.sqlalchemy.sqla.types import BigInteger
 from graphene.utils.subclass_with_meta import (
     SubclassWithMeta_Meta as _BaseObjectTypeMetaclass)
 from graphene_sqlalchemy import SQLAlchemyObjectType as _SQLAObjectType
+from graphene_sqlalchemy.converter import (
+    convert_sqlalchemy_type, convert_column_to_int_or_id, get_column_doc, is_column_nullable)
 from graphene_sqlalchemy.types import (
     SQLAlchemyObjectTypeOptions as _SQLAObjectTypeOptions)
+from sqlalchemy import types
 from sqlalchemy.orm import class_mapper
+
+
+# convert_sqlalchemy_type.register(BigInteger)(convert_column_to_int_or_id)
+# convert_sqlalchemy_type.register(types.BigInteger)(convert_column_to_int_or_id)
+
+@convert_sqlalchemy_type.register(BigInteger)
+@convert_sqlalchemy_type.register(types.BigInteger)
+def convert_column_to_int_or_id(type, column, registry=None):
+    if column.primary_key or column.foreign_keys:
+        return graphene.ID(
+            description=get_column_doc(column),
+            required=not (is_column_nullable(column)),
+        )
+    else:
+        return graphene.Int(
+            description=get_column_doc(column),
+            required=not (is_column_nullable(column)),
+        )
 
 
 class SQLAlchemyObjectTypeOptions(_SQLAObjectTypeOptions):
@@ -98,7 +120,7 @@ class SQLAlchemyObjectType(_SQLAObjectType):
         if unchained._models_initialized:
             model = unchained.sqlalchemy_bundle.models[model.__name__]
 
-            # graphene has a horrible habbit of eating exceptions and this is one
+            # graphene has a horrible habit of eating exceptions and this is one
             # place where it does, so we preempt it (if this fails it should throw)
             class_mapper(model)